Alpha???
Linked the frontend with backend. Created various functions to handle events. Needs CSS for new HTML I created. So far have basic functionality (purchase multiple ventures and upgrade them). Some of the venture properties need tweeking as the game feels a little weird (too easy or no challenge), but this was to be expected from the current state. Still need to work on the time counter at the top right. Information about ventures that is displayed needs a little work as if feels very empty at the moment.